requirements.txt
aiogram==3.15.0
python-dotenv==1.0.1
loguru==0.7.3
flake8==7.1.1
apscheduler==3.11.0
psycopg2==2.9.10